requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-API==1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pipenv==2018.11.26
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
slackclient==1.3.1
urllib3==1.24.2
virtualenv==16.4.3
virtualenv-clone==0.5.1
websocket-client==0.54.0
Werkzeug==0.15.3
gunicorn==19.9.0
